refactor: color.enum에서 place.color.enum으로 네이밍 변경 #117
