+++ backend/src/map/dto/AddPlaceToMapRequest.ts
@@ -1,5 +1,5 @@
 import { IsNumber, IsString, IsEnum } from 'class-validator';
-import { Color } from '../../place/color.enum';
+import { Color } from '../../place/place.color.enum';
 
 export class AddPlaceToMapRequest {
   @IsNumber()
